libusbg: Make *_attrs and *_strs arguments const-correct
This makes it easier to use libusbg in clean, idiomatic C++ code.

This patch changes the *_attrs and *_strs arguments only. It does not
make any assumptions regarding the const-ness of other structures
(usbg_gadget etc.), which is something that doesn't come up in typical
use cases.

Signed-off-by: Dominic Sacré <dominic.sacre@gmx.de>

Add libusbg: to commit message.

Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>